pyproject.toml
[project]
name = "docmind-rag"
version = "0.1.0"
description = "Document Q&A with RAG"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain<0.4.0",
    "langchain-community<0.4.0",
    "langchain-google-genai",
    "langchain-chroma",
    "langchain-huggingface",
    "langchain-text-splitters<0.4.0",
    "chromadb",
    "pypdf",
    "python-dotenv",
    "sentence-transformers>=5.2.2",
    "google-generativeai>=0.8.6",
    "pymupdf>=1.27.1",
]
